code clean up and  added readme
